requirements.txt
bcrypt==4.0.1
cachelib==0.10.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
flask-mongoengine==1.0.0
Flask-Session==0.4.0
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.27.0
Pillow==9.5.0
pymongo==4.3.3
python-dotenv==1.0.0
Werkzeug==2.2.3
WTForms==3.0.1
